Fix rinning of docker image
It is not allowed and not necessary --user parametr in entrypoint because of already applied USER directive earlier in dockerfile.
Moreover with --user in entrypoint builded docker image does not work (at least in debian).
we can appply --user parametr only if USER root directive used, but it is not conformant to openshift container certification. For USER maxscale we get access denied. (maxscale user does not have such provilege)