Fix running of docker container
It is not allowed and not necessary --user parametr in entrypoint because of already applied USER directive earlier in dockerfile. Moreover with --user in entrypoint builded docker image does not work (at least in debian). we can apply --user parametr only if USER root directive used, but it is not conformant to openshift container certification. For USER maxscale we get access denied. (maxscale user does not have such provilege)

I am contributing the new code of the whole pull request, including one or
several files that are either new files or modified ones, under the BSD-new
license.